+++ database/migrations/2026_05_10_051504_create_order_items_table.php
@@ -11,16 +11,18 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('order_items', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('order_id')->constrained()->onDelete('cascade');
-            $table->foreignId('product_id')->nullable()->constrained()->onDelete('set null');
-            $table->string('product_name');
-            $table->decimal('price', 10, 2);
-            $table->integer('qty');
-            $table->decimal('total', 10, 2);
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('order_items')) {
+            Schema::create('order_items', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('order_id')->constrained()->onDelete('cascade');
+                $table->foreignId('product_id')->nullable()->constrained()->onDelete('set null');
+                $table->string('product_name');
+                $table->decimal('price', 10, 2);
+                $table->integer('qty');
+                $table->decimal('total', 10, 2);
+                $table->timestamps();
+            });
+        }
     }
 
     /**
